+++ src/pages/ManageCommunities/ManageCommunities.jsx
@@ -40,15 +40,20 @@ function ManageCommunities() {
     };
   
     const handleDeleteClick = async (id) => {
-      try {
-        const response = await axios.delete(FETCH_ALL_COMMUNITIES+`/${id}`, { withCredentials: true });
-        if (response.data.success) {
-          setCommunities(communities.filter(community => community.id !== id));
-        } else {
-          console.error("Error deleting property:", response.data.message);
+      const confirmDelete = window.confirm(
+        "Are you sure you want to delete this community?"
+      );
+      if (confirmDelete) {
+        try {
+          const response = await axios.delete(FETCH_ALL_COMMUNITIES+`/${id}`, { withCredentials: true });
+          if (response.data.success) {
+            setCommunities(communities.filter(community => community.id !== id));
+          } else {
+            console.error("Error deleting community:", response.data.message);
+          }
+        } catch (error) {
+          console.error("Error deleting community:", error);
         }
-      } catch (error) {
-        console.error("Error deleting property:", error);
       }
     };
